fix(logic): Update ByteDocument and LineIndexedDocument
This commit fixes an issue in ByteDocument.cpp and LineIndexedDocument.cpp.
In ByteDocument.cpp, the assert statement now includes a check for position
being greater than or equal to m_BOMsize. In LineIndexedDocument.cpp, the
eol variable is now correctly assigned based on whether the character at the
end position is a newline character. (main)