feat(textviewers): Add setKind() method to TextViewer
This commit adds a new method setKind() to the TextViewer class. The
method allows the user to set the kind of logic for the PaintArea
widget. The kind parameter specifies the type of logic to be used.
The method calls the setKind() method of the PaintArea widget and
updates the logicKind variable accordingly. The setData() method is
then called to update the data based on the new logicKind value.

Issue: main